Guarantee lock freeing using RAII.
Calling unlock() outside of a finally block is almost always
a source of bugs/deadlocks. If the Xerces-c initialization
throws an exception the mutex would never get unlocked.
Using std::lock_guard to lock the mutex will automatically
unlock the mutex (with or without) exception when the block
is exited.