Fix upper-/lowercase node names.
Now that attributes are returned with correct
casing, we need to adapt the node names to
contain correct casing as well.
This isn't always possible as some node names
can occurr in upper- and lowercase versions
depending on the context they are in.

As the matching currently doesn't check the
casing, there is no reason to introduce
duplicate node types for these. But that
would be the correct way to handle it.